feat: routers completos para categories, products, orders y stats

+++ app/schemas/stats.py
@@ -1,5 +1,5 @@
 # app/schemas/stats.py
-from pydantic import BaseModel
+from pydantic import BaseModel, field_serializer
 from typing import List
 from decimal import Decimal
 
@@ -8,18 +8,18 @@ class TopProduct(BaseModel):
     product_id: int
     product_name: str
     total_sold: int
-    total_revenue: Decimal
+    total_revenue: float
 
 
 class DailySales(BaseModel):
     date: str
     total_orders: int
-    total_revenue: Decimal
+    total_revenue: float
 
 
 class StatsResponse(BaseModel):
-    total_revenue: Decimal
+    total_revenue: float
     total_orders: int
-    average_ticket: Decimal
+    average_ticket: float
     top_products: List[TopProduct]
     daily_sales: List[DailySales]
